Changes the way users are compared when filtering
When filtering by profession, the comparison was made by objects
(because the value of the 'profession' field the user had an object with
a profession.
Now in fake api professions represent 2 different objects, and when
compared they are always not equal to each other. Therefore, the
comparison now takes place by the value of the '_id' field, in which a
unique number is represented in string form.