Refactoring
The addSortMarker function has been changed - previously, a variable
with the base part of the class for the marker element was created in
any case. But it was used only if you need to add a sorting icon in the
column header. Now the variable containing the class creates only for
those columns for which sorting is provided.

Previously, the table itself as a composition of components could be
reused only if the data necessary for drawing the table was transmitted
to the Table component via the passes.
Now, to reuse the table, you can not only transfer data through the
props (leaving the rendering of the TableHeader and TableBody components
to the Table component), but also transfer the TableHeader and TableBody
components as child components for the Table~