Changes the logic of rendering components
When you need to draw the 'Users' page during routing, different
components are drawn inside depending on the optional parameters.

If only the user ID is passed, then the user card is drawn.

If, in addition to the user id, there is also the 'edit' parameter, then
the component with the edit page is drawn.